Add lineCounts(player)  myTurn()   squares(player)
All as part of working on getBoardHints().  Next up is completing  listThrees(player) and  listTwos(player)  and  listOnes(player)  which respectively filter the array result of lineCounts() according to how many of that palyers piece are in that line ONLY if the remaining squares in that line are all EMPTY.  I could have made a distinction like  listTwos(player) listing ALL two in a rows and listUnblockedTwos(player) doing that but for now I see no use for a method that includes blocked one and two in a rows.