Add default prop to most methods (moveList = history)
Most methods either used the history state directly or depended on other methods that did.  This let me write very neat looking code since most of my functions required no parameters.  Later on I wanted to add the ability to use any of these methods on a hypothetical version of the history to test future possible positions.  This hypothetical moveList can be passed in and if it is not it defaults to using history.