simplify doubleAttackCreatingMoves()
getBoardHints has steps 1 and 2 that use immediateWins() and urgentDefensiveMoves() and several steps after that.  getBoardHints() became bloated with logic about what square colors to assign based on calls to methods that may return overlapping sets.  Removed some complexity by adding EARLY RETURNS to step 1 and 2 so that step 3 doesn't have to begin with a complex condition that refers back to what modifications to the hints array 1 and 2 already made.