Found VERY sneaky bug!
in the body of linesWithOnlyOne(moveList)  there was a call to  lineCountsFor(other(player, moveList))  This is a TYPO and should say lineCountsFor(other(player), moveList)    lineCountsFor takes two props, a player and a move list. they other(player) helper does not need the move list passed to it.  Part of why this was so difficult to find is that it was done correctly in  linesWithOnlyTwo()  and I originally wrote  linesWithOnlyOne() long before I had a need for it.  Since it was writeen so long ago and not changed I assumed it ws correct but only  linesWithOnlyTwo() had been put to use to prove it worked as expected in play when lines can be blocked.