Fixed a Super Sneaky Bug!
The function immediateWins(moveList = history) was originally written with no parameters and only worked with the history but it was later given a default param.  Originally it assigned const player = myTurn() and history was used automatically!!  It called myTurn() without passing it the moveList.  So when immediateWins() was invoked by thereIsAForcedWin() it gave an innacurate result because it was saying "Yes there are immediate wins" when it was mistaken about whose turn it was!