Add steps 4-8 to getBoardHints
Almost working but there is an inaccuracy in how doubleAttackCreatingMoves works that caused it to ignore double attacks that involve a pre-existing threat.  It is plausible to ignore an immediateWin and still win in spite of that... this should be reflected in the hints even though it involves "non-optimal" play.