Resolved gameOver wins() bug.
There was a difficult bug where the wins() method accepted moveList as an optional second parameter to "player" but this defaulted to history.
Inside handleCardClick() I got an incorrect result the first time gameOver() was called, resulting in the Bot making a move after the player has actually won.  Oddly the calls to gameOver() that gave data to the Panel were working because by the time they were made the history update had completed.  To resolve this I needed to explicitly pass the moveList to both calls that gameOver makes to wins().