Work on How To Get a Best Move.
I really want a way to define this process reecursively rather than step by step as I did in the Tic-Tac-Toe code.
We will create a QUEUE of the positions that need to be checked. We will use this structure to facilitate a breadth-first-search for an immediatelyWinningMove()   it progressively filters a list of "possibleWinningMoves()" that starts as all unclaimedNumbers() .
If there are no winning moves in the current position then look for losing moves and filter the possibleWinningMoves() to not include any definite losing move.