Split into two pages
Rather than having switches in a settings panel for commentary and hints this is being broken into two separate routes that render the tic tac toe page with a new prop - "mode" which can be "learn" or "play".  Will soon add a button that shows hints on board one turn at a time to the learn panel.